Added ticker
Here are a few notes which should cover the gist of the changes:

The calendar was being read every second which turned out to be a bad things for a few reasons
to change that i made messages a global variable @ line 186 and getMsg() no longer calls updateMsg, is only called once at init

updateMsg() is a thing now

in getMsg(), i appended two boolean values to each element of messages, they're used in updateMsg.
there might be a better place for that other than in getMsg(), i'll admit

added a lot of css

i had a hard problem with clipPath that i wasnt able to solve, tried using <pattern> instead but i suck at that, so i used a mask in the end. it punches a hole inside the main big black circle to see the display window. so the svg stuff changed quite a bit