Increased ticker animation flexibility
Tiny change, the string slice at line 79 now removes the last character (the time unit 's') instead of only keeping the first one (so a duration of "2.3s" actually returns 2.3). This allows us to use multiple digits when setting the duration of the ticker's message transition in the css.

Also deleted a bunch of testing stuff i forgot to remove at the end of updateMsg(msgJSON).